95fcoe: Do not overwrite FCoE configuration
If the installation already has a FCoE configuration we should
not attempt to overwrite it but rather use the pre-defined
configuration.

References: bsc#993861

+++ modules.d/95fcoe/fcoe-up.sh
@@ -38,6 +38,7 @@ netdriver=$(readlink -f /sys/class/net/$netif/device/driver)
 netdriver=${netdriver##*/}
 
 write_fcoemon_cfg() {
+    [ -f /etc/fcoe/cfg-$netif ] && return
     echo FCOE_ENABLE=\"yes\" > /etc/fcoe/cfg-$netif
     if [ "$dcb" = "dcb" ]; then
         echo DCB_REQUIRED=\"yes\" >> /etc/fcoe/cfg-$netif
